[#81] : Code cleanup

+++ src/material_manager.cpp
@@ -11,7 +11,7 @@ manta::MaterialManager::~MaterialManager() {
 }
 
 manta::Material *manta::MaterialManager::searchByName(const std::string &name) const {
-	int nMaterials = m_materials.size();
+	int nMaterials = (int)m_materials.size();
 
 	for (int i = 0; i < nMaterials; i++) {
 		if (m_materials[i]->getName() == name) {
